Updated the calaculate button in main page

+++ lib/formula_bmi.dart
@@ -1,7 +1,7 @@
 import 'dart:math';
 
 class formula {
-  double bmi_calulation(double height, double weight) {
+  double bmi_calulation(int height, int weight) {
     return (weight / pow(height, 2));
   }
 }
